move Wallet.tsx to types
