+++ wallet/types/Wallet.tsx
